Change project inference interface
Accept glob patterns when inferring projects so certain patterns can
be excluded, e.g., when a project contains example code that is not
part of the compiled project.

Throw an exception when a project has more than one Spring Boot
application annotation instead of silently ignoring it.

Move reused exports from the Spring Boot project structure tests to
another file so those tests do not get run twice.
