+++ test/spring/transform/addAnnotationToSpringBootClass.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018 Atomist, Inc.
+ * Copyright © 2019 Atomist, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import { addAnnotationToSpringBootClass } from "../../../lib/spring/transform/ad
 import {
     GishJavaPath,
     GishProject,
-} from "../generator/SpringBootProjectStructure.test";
+} from "../generator/springProjects";
 
 describe("addAnnotationToSpringBootClass", () => {
